[ui] Add main and login routes
Adds vue-router as a dependency to manage routes.
The main dashboard view is moved to its own component
and is displayed when authenticaded on the '/' route.
If a user tries to access that route unauthenticated, they
are redirected to '/login'.
A button with the logged username is displayed on the app
bar that allows the user to log out. The token and user
cookies are removed when logged out.
