+++ ui/package.json
@@ -20,7 +20,8 @@
     "graphql-tag": "^2.10.3",
     "js-cookie": "^2.2.1",
     "vue": "^2.6.11",
-    "vue-apollo": "^3.0.3", 
+    "vue-apollo": "^3.0.3",
+    "vue-router": "^3.4.9",
     "vuetify": "^2.2.11",
     "vuex": "^3.1.3"
   },
