+++ ui/src/router/index.js
@@ -1,0 +1,40 @@
+import Router from "vue-router";
+import store from "../store";
+import Dashboard from "../components/Dashboard";
+import Login from "../components/Login";
+
+const routes = [
+  {
+    path: "/",
+    name: "Dashboard",
+    component: Dashboard,
+    meta: { requiresAuth: true }
+  },
+  {
+    path: "/login",
+    name: "Login",
+    component: Login
+  }
+];
+
+const router = new Router({
+  mode: "history",
+  routes
+});
+
+router.beforeEach((to, from, next) => {
+  const isAuthenticated = store.getters.isAuthenticated;
+  if (to.matched.some(record => record.meta.requiresAuth)) {
+    if (!isAuthenticated) {
+      next({
+        path: "/login"
+      });
+    } else {
+      next();
+    }
+  } else {
+    next();
+  }
+});
+
+export default router;
